acrentado valores para fazer calculos desejados mudança de valor x de inteiro pra double

+++ JavaApp/src/javaapp/JavaApp.java
@@ -6,7 +6,7 @@ public class JavaApp {
 
    
     public static void main(String[] args) {
-        int x;
+        double x;
         double y;
             System.out.println(" digite a pimeira nota : ");
